Rename name space to Microsoft.DotNet.ToolPackage
