+++ src/dotnet/ToolPackage/ToolConfigurationDeserialization/DotNetCliTool.cs
@@ -1,7 +1,7 @@
 ﻿using System.Diagnostics;
 using System.Xml.Serialization;
 
-namespace Microsoft.DotNet.ToolPackageObtainer.ToolConfigurationDeserialization
+namespace Microsoft.DotNet.ToolPackage.ToolConfigurationDeserialization
 {
     [DebuggerStepThrough]
     [XmlRoot(Namespace = "", IsNullable = false)]
