+++ test/Microsoft.DotNet.ShellShim.Tests/FakeEnvironmentProvider.cs
