+++ test/Microsoft.DotNet.ShellShim.Tests/FakeFile.cs
