+++ test/Microsoft.DotNet.ShellShim.Tests/FakeReporter.cs
