+++ test/Microsoft.DotNet.ShellShim.Tests/LinuxEnvironmentPathTests.cs
