+++ test/Microsoft.DotNet.ShellShim.Tests/OsxEnvironmentPathTests.cs
