+++ test/Microsoft.DotNet.ShellShim.Tests/ShellShimMakerTests.cs
