+++ test/Microsoft.DotNet.ToolPackage.Tests/DotnetToolSettingsGolden.xml
