+++ test/Microsoft.DotNet.ToolPackage.Tests/DotnetToolSettingsMalformed.xml
