+++ test/Microsoft.DotNet.ToolPackage.Tests/DotnetToolSettingsMissing.xml
