+++ test/Microsoft.DotNet.ToolPackage.Tests/SampleGlobalTool/Program.cs
