+++ test/Microsoft.DotNet.ToolPackage.Tests/SampleGlobalTool/consoledemo.csproj
