+++ test/Microsoft.DotNet.ToolPackage.Tests/SampleGlobalTool/includepublish.nuspec
